requirements.txt
python-binance==1.0.12
python-telegram-bot==13.7
gunicorn
requests

